chore: Move fixtures to conftest to reuse.
